requirements.txt
fpdf==1.7.2
python-docx==0.8.11
pdfkit==1.0.0
streamlit==1.10.0
langdetect==1.0.9
odfpy==1.4.1
altair<5
markdown==3.6